Added dropdown options for nature/demeanor
Nature & Demeanor typically have a list of set options.
I chose to leave it more free-form with text boxes. However, each
defined nature & demeanor in the rulebook has a set effect on
gameplay/roleplay. I didn't realize this. Replacing the textboxes with
dropdown makes any custom nature/demeanor saved to a character unusable.
So, since I need to keep the textbox and add a dropdown, this was the
best solution.